- fixed module dependency retrieval - refactored the files frontend index controller to use the new request flow - fixed the Image class

+++ ACP3/Core/Cache.php
@@ -50,6 +50,10 @@ class Cache
      */
     public static function purge($dir, $cacheId = '')
     {
+        if (!is_file($dir) && !is_dir($dir)) {
+            return true;
+        }
+
         if (is_file($dir) === true) {
             return @unlink($dir);
         }
